+++ src/main/kotlin/next/step/minesweeper/controller/MineSweeperController.kt
@@ -1,11 +1,11 @@
 package next.step.minesweeper.controller
 
-import next.step.blackjack.view.OutputView
 import next.step.minesweeper.domain.board.Board
 import next.step.minesweeper.domain.board.BoardArea
 import next.step.minesweeper.domain.mine.generator.RandomMineGenerator
 import next.step.minesweeper.utils.retryOnFailure
 import next.step.minesweeper.view.InputView
+import next.step.minesweeper.view.OutputView
 
 fun main() = retryOnFailure {
     val boardArea = BoardArea(InputView.readHeight(), InputView.readWidth())
@@ -18,7 +18,7 @@ fun main() = retryOnFailure {
         { OutputView.showError(it.message) },
         { OutputView.showBoardPoints(it) },
         { OutputView.showSuccess() },
-        { OutputView.showFail() }
+        { OutputView.showFail() },
     )
 }
